finish up caching, make aoc6 pt2 work
the part that was weird for me was that if you did a get() on the
cache, you could not also modify it without being done with the value
you did a get on, since the value ends up causing an immutable borrow.
I worked around it by using remove instead, which moved the value out
of the cache and free it. But then again, I ended up using a usize
cache, which implements clone, so there was no such problem in the end.